requirements.txt
streamlit
numpy
pandas


# Face - recognition 

scikit-learn
dlib-bin

git+https://github.com/ageitgey/face_recognition_models
setuptools<70.0.0

supabase
bcrypt

segno
pillow



librosa
resemblyzer